requirements.txt
addict==2.4.0
alembic==1.4.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
commonmark==0.9.1
cycler==0.10.0
Cython==0.29.23
databricks-cli==0.14.3
decorator==4.4.2
decord==0.4.1
defusedxml==0.7.1
docker==5.0.0
einops==0.3.0
entrypoints==0.3
ez-setup==0.9
ffmpeg-python==0.2.0
flake8==4.0.1
Flask==2.0.0
future==0.18.2
gitdb==4.0.7
GitPython==3.1.17
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
importlib-metadata==4.8.2
itsdangerous==2.0.0
Jinja2==3.0.0
kiwisolver==1.3.1
Mako==1.1.4
Markdown==3.3.6
MarkupSafe==2.0.0
matplotlib==3.4.2
mccabe==0.6.1
mlflow==1.17.0
mmaction2==0.20.0
mmcv-full==1.3.18
mmdet==2.12.0
model-index==0.1.11
moviepy==1.0.3
numpy==1.20.2
odfpy==1.4.1
opencv-contrib-python==4.5.4.60
opencv-python==4.5.4.60
openmim==0.1.5
ordered-set==4.0.2
packaging==21.3
pandas==1.1.5
Pillow==8.2.0
proglog==0.1.9
prometheus-client==0.10.1
prometheus-flask-exporter==0.18.2
protobuf==3.17.0
pycocotools==2.0.2
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
querystring-parser==1.2.4
requests==2.25.1
rich==10.9.0
scipy==1.6.3
seaborn==0.11.1
six==1.16.0
smmap==4.0.0
soupsieve==2.2.1
SQLAlchemy==1.4.15
sqlparse==0.4.1
tabulate==0.8.9
terminaltables==3.1.0
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchvision==0.11.1+cu113
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.4
websocket-client==0.59.0
Werkzeug==2.0.0
yapf==0.31.0
zipp==3.4.1
